Fixed a couple of omissions

+++ scripts/mailpile-admin.py
@@ -46,7 +46,7 @@ MAILPILE_DELETE_SCRIPT = [
 
 INSTALL_APACHE_SCRIPT = [
     '"%(packager)s" install screen expect',
-    'a2enmod headers rewrite proxy proxy_http',
+    'a2enmod headers rewrite proxy proxy_http cgi',
     'mkdir -p /var/lib/mailpile/apache/ /var/lib/mailpile/pids/',
     'cp -a "%(mailpile-www)s"/* /var/lib/mailpile/apache/',
     'rm -f /var/lib/mailpile/apache/shared',
@@ -290,7 +290,8 @@ def save_htaccess(args, os_settings, mailpiles):
     with open(APACHE_HTACCESS_PATH + '.new', 'w') as fd:
         os_settings['rewriterules'] = '\n'.join(rules)
         fd.write(APACHE_HTACCESS_TEMPLATE % os_settings)
-    os.remove(APACHE_HTACCESS_PATH)
+    if os.path.exists(APACHE_HTACCESS_PATH):
+        os.remove(APACHE_HTACCESS_PATH)
     os.rename(APACHE_HTACCESS_PATH + '.new', APACHE_HTACCESS_PATH)
